Add files via upload
These notebooks are more experimental and for purpose in experiments without grand design. 
They are more messy accordingly due to their one time use intent.

The detail exploration of simulations around a poitnwhere the lorenz is born, and dies.

A poor mans continuation that is computationally infeasible due to an extensive list of attractor duplicates

A notebook to analyse the flow of confabulations and their relationship to the Lorenz attractors flow

And a notebook that was used to test ideas before commiting them to cleaner notebooks.